backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.9
python-dotenv==1.1.1
openai==1.109.1
pypdf==5.9.0
reportlab==4.2.2
